# src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Travel Planner API",
    "version": "1.0.0",
    "description": "Classroom API with open CORS, 7-digit matricula auth header, and admin/student roles."
  },
  "servers": [
    {
      "url": "https://leiame.app",
      "description": "Production"
    }
  ],
  "security": [
    {
      "MatriculaAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "MatriculaAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Use a 7-digit matricula for student requests or the admin token from config/admin.json to access admin routes."
      }
    },
    "schemas": {
...
      "Location": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "summary": "Echo bearer token for a matricula",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "matricula": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token issued"
          }
        },
        "operationId": "login"
      }
    },
    "/locations": {
      "get": {
        "summary": "List locations (cities/regions)",
        "responses": {
          "200": {
            "description": "Locations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Location"
                  }
                }
              }
            }
          }
        },
        "operationId": "listLocations"
      },
      "post": {
        "summary": "Create location (admin token required)",
        "security": [
          {
            "MatriculaAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "operationId": "createLocation"
      }
    },
    "/locations/{id}": {
      "get": {
        "summary": "Get location by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Location"
          },
          "404": {
            "description": "Not found"
          }
        },
        "operationId": "getLocation"
      },
      "put": {
        "summary": "Update location (admin token required)",
        "security": [
          {
            "MatriculaAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated"
          },
          "404": {
            "description": "Not found"
          }
        },
        "operationId": "updateLocation"
      },
      "delete": {
        "summary": "Delete location (admin token required)",
        "security": [
          {
            "MatriculaAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          },
          "404": {
            "description": "Not found"
          }
        },
        "operationId": "deleteLocation"
      }
    },
    "/hotels": {
      "get": {
        "summary": "List hotels",
        "parameters": [
          {
            "in": "query",
            "name": "city",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "priceMin",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "priceMax",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "stars",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "amenities",
            "schema": {
              "type": "string",
              "description": "Comma-separated list, e.g. wifi,pool"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hotel list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Hotel"
                  }
                }
              }
            }
          }
        },
        "operationId": "listHotels"
      },
      "post": {
        "summary": "Create hotel (admin)",
        "security": [
          {
            "MatriculaAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "operationId": "createHotel"
      }
    },
    "/hotels/{id}": {
      "get": {
        "summary": "Hotel by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hotel"
          },
          "404": {
            "description": "Not found"
          }
        },
        "operationId": "getHotel"
      }
    },
    "/hotels/{id}/availability": {
      "get": {
        "summary": "Hotel availability",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Availability"
          }
        },
        "operationId": "getHotelAvailability"
      }
    },
    "/hotels/{id}/reviews": {
      "get": {
        "summary": "List reviews for a hotel",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reviews",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Review"
                  }
                }
              }
            }
          }
        },
        "operationId": "listHotelReviews"
      },
      "post": {
        "summary": "Add review for a hotel (matrícula token required)",
        "security": [
          {
            "MatriculaAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rating": {
                    "type": "integer"
                  },
                  "comment": {
                    "type": "string"
                  }
                },
                "required": [
                  "rating"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "operationId": "createHotelReview"
      }
    },
    "/planes": {
      "get": {
        "summary": "List flights",
        "parameters": [
          {
            "in": "query",
            "name": "origin",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "destination",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Flights",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Plane"
                  }
                }
              }
            }
          }
        },
        "operationId": "listPlanes"
      },
      "post": {
        "summary": "Create flight (admin)",
        "security": [
          {
            "MatriculaAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "operationId": "createPlane"
      }
    },
    "/planes/search": {
      "get": {
        "summary": "Search flights (alias of /planes)",
        "parameters": [
          {
            "in": "query",
            "name": "origin",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "destination",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Flights",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Plane"
                  }
                }
              }
            }
          }
        },
        "operationId": "searchPlanes"
      }
    },
...
    "/purchases": {
      "post": {
        "summary": "Create purchase (self or admin)",
        "security": [
          {
            "MatriculaAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "operationId": "createPurchase"
      }
    },
...
  }
}